Move paginate_cursor to a concern
